AIX_ML_Models/aix_reconnaissance/requirements.txt
sahi 
yolov5==7.0.4
opencv-python
fastapi
python-multipart
uvicorn
Pillow
requests
numpy
kafka-python